pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice-controlled local AI agent with FastAPI backend and browser frontend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
    "groq==0.13.1",
    "langchain==0.3.14",
    "langchain-groq==0.2.4",
    "langgraph==0.2.73",
    "pydantic==2.10.4",
    "httpx==0.28.1",
    "aiofiles==24.1.0",
]
